--- a/src/native/QmlNet/QmlNet/types/NetReference.h
+++ b/src/native/QmlNet/QmlNet/types/NetReference.h
@@ -6,15 +6,13 @@
 class NetReference
 {
 public:
-    NetReference(NetGCHandle* gcHandle, uint64_t objectId, QSharedPointer<NetTypeInfo> typeInfo);
+    NetReference(uint64_t objectId, QSharedPointer<NetTypeInfo> typeInfo);
     ~NetReference();
-    NetGCHandle* getGCHandle();
     uint64_t getObjectId();
     QSharedPointer<NetTypeInfo> getTypeInfo();
 
     void release();
 private:
-    NetGCHandle* gcHandle;
     uint64_t objectId;
     QSharedPointer<NetTypeInfo> typeInfo;
 };