--- a/src/native/QmlNet/QmlNet/types/NetReference.h
+++ b/src/native/QmlNet/QmlNet/types/NetReference.h
@@ -10,8 +10,6 @@ public:
     ~NetReference();
     uint64_t getObjectId();
     QSharedPointer<NetTypeInfo> getTypeInfo();
-
-    void release();
 private:
     uint64_t objectId;
     QSharedPointer<NetTypeInfo> typeInfo;