References are counted manually on .Net side
Behaviors now don't get an event on each NetReference create / delete but when an .Net object enters Native the first time and when it leaves the Native space