@devmil 2018-07-24
Seems this change introduced an unpredictable side effect. Each run some other tests are failing. Sometimes none. Not good...