@devmil 2018-07-24
The ObjectTagger unit tests reset the object id database and introduce this side effect.
As we now have the reverse lookup and use this to retrieve the object instances this problem has been made visible