@pauldotknopf 2018-07-24
# With GCHandle allocations

``` ini

BenchmarkDotNet=v0.11.0, OS=macOS High Sierra 10.13.6 (17G65) [Darwin 17.7.0]
Intel Core i7-3635QM CPU 2.40GHz (Ivy Bridge), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=2.1.301
  [Host] : .NET Core 2.1.1 (CoreCLR 4.6.26606.02, CoreFX 4.6.26606.05), 64bit RyuJIT
  Core   : .NET Core 2.1.1 (CoreCLR 4.6.26606.02, CoreFX 4.6.26606.05), 64bit RyuJIT

Job=Core  Runtime=Core  Force=True  
InvocationCount=1  UnrollFactor=1  

```
| Method |    Mean |    Error |   StdDev |      Gen 0 | Allocated |
|------- |--------:|---------:|---------:|-----------:|----------:|
|    Run | 8.534 s | 0.2023 s | 0.5868 s | 40000.0000 | 112.38 MB |

# Without GCHandle allocations

``` ini

BenchmarkDotNet=v0.11.0, OS=macOS High Sierra 10.13.6 (17G65) [Darwin 17.7.0]
Intel Core i7-3635QM CPU 2.40GHz (Ivy Bridge), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=2.1.301
  [Host] : .NET Core 2.1.1 (CoreCLR 4.6.26606.02, CoreFX 4.6.26606.05), 64bit RyuJIT
  Core   : .NET Core 2.1.1 (CoreCLR 4.6.26606.02, CoreFX 4.6.26606.05), 64bit RyuJIT

Job=Core  Runtime=Core  Force=True  
InvocationCount=1  UnrollFactor=1  

```
| Method |    Mean |    Error |   StdDev |      Gen 0 | Allocated |
|------- |--------:|---------:|---------:|-----------:|----------:|
|    Run | 6.973 s | 0.0800 s | 0.0749 s | 34000.0000 | 104.68 MB |

# Comments

The overall speed is 23% faster! I was expected more, but that works! I think even better is the standard deviation. The old way deviated about a .58 seconds from the mean, while the new way only deviated ~.07. To me, that tells me that garbage collection is a lot more efficient and deterministic.

Merging!